Divide OpenReacRunnerTest into several TUs
